% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpenwls.R
\name{dcpenwls}
\alias{dcpenwls}
\title{Fit the doubly interval-censored AFT model with linear regression model}
\usage{
dcpenwls(
  L,
  R,
  T,
  delta,
  x,
  estimation = NULL,
  beta0,
  type,
  wttype = "KM",
  hlimit = 0.1,
  id = NULL,
  index = NULL,
  nonzero.index = NULL,
  lamb.len = 200,
  maxit = 100,
  tol = 1e-10
)
}
\arguments{
\item{L}{left-censoring time, having 0 if left-censored.}

\item{R}{right-censoring time, having \code{Inf} if right-censored.}

\item{T}{exactly observed time.}

\item{delta}{censoring indicator, 1: exactly observed; 2: left-censored; 3: right-censored;}

\item{x}{X matrix of baseline covariates.}

\item{estimation}{estimating method of partly interval censored, if estimation="DR", doubly robust estimator is estimated.}

\item{type}{penalized estimating method, default is "wls", lasso penalized estimating method as "lasso", elasticnet penalized estimating method as "elasticnet", SCAD penalized estimating method as "SCAD"}

\item{wttype}{weight estimating method, default is "KM", Beran's nonparametric KM estimating method as "Beran", and  Ishwaran's random survival forests KM estimating method as "Ishwaran".}

\item{hlimit}{bandwidth value, default is NULL.}

\item{id}{cluster id. If the data does not have clustered structure, set \code{id=NULL}.}

\item{index}{index of cluster weight, default is 1}

\item{nonzero.index}{index of nonzero position, default is NULL.}

\item{lamb.len}{the length of grid of lambdas, default is NULL.}

\item{maxit}{maximum number of iteration for the log-rank estimator, default is 100.}

\item{tol}{tolerance of iteration for the log-rank estimator, default is 1e-3.}
}
\value{
\code{dcpenwls} returns a data frame containing at least the following components:
\itemize{
\item \code{coefficients}: regression estimator.
\item \code{se}: standard error estimates for \code{est}.
\item \code{pvalue}: p-value.
\item \code{lower bd}: lower bound of coefficients under 95\% confidence level.
\item \code{upper bd}: upper bound of coefficients under 95\% confidence level.
}
}
\description{
Fit inverse weighted linear regression with doubly interval-censored data
}
\details{
see Kim et al., (2023+) for detailed method explanation.
}
\examples{
\dontrun{
# Simulations
n=200; x1=runif(n,-1.2,1.7); x2=rbinom(n,1,0.6)
T = 1.7+x1+x2+rnorm(n)*(1-0.1*x2)
L=runif(n,-2.8,1.9); R=L+runif(n,4.2,8.1)
Y=pmin(R,pmax(T,L))
delta=case_when(
 T<L ~ 3,
 T>R ~ 2,
 TRUE ~ 1 #observed
)
L=L; R=R; T=T; delta=delta; x=cbind(x1,x2); tau=0.3
dcwls(L,R,T,delta,x)
dcpenwls(L,R,T,delta,x=x,hlimit=0.1,wttype="KM")
}
}
\references{
Beran, R. (1981). Nonparametric Regression with Randomly Censored Survival Data. Technical Report, Univ.California, Berkeley.

Ishwaran, H., U. B. Kogalur, E. H. Blackstone, and M. S. Lauer (2008). Random survival forests. The annals of applied statistics. 2 (3), 841–860.

Kim, Y., Choi, S. (2023+). On weighted-least squares regression with partially interval-censored data.
}
